+++ extension.json
@@ -28,8 +28,8 @@
 		"LoopUpdater": "schema/LoopUpdater.php",
 		"LoopStructure": "includes/LoopStructure.php",
 		"LoopStructureItem": "includes/LoopStructure.php",
-		"SpecialLoopStructure": "includes/SpecialLoopStructure.php",
-		"SpecialLoopStructureEdit": "includes/SpecialLoopStructure.php",
+		"SpecialLoopStructure": "includes/LoopStructure.php",
+		"SpecialLoopStructureEdit": "includes/LoopStructure.php",
 		"LoopXml": "includes/LoopXml.php",
 		"LoopPdf": "includes/LoopPdf.php",
 		"LoopMp3": "includes/LoopMp3.php",
@@ -43,10 +43,10 @@
 		"LoopExportEpub": "includes/LoopExport.php",
 		"LoopExportHtml": "includes/LoopExport.php",
 		"LoopExportScorm": "includes/LoopExport.php",
-		"SpecialLoopExport": "includes/SpecialLoopExport.php",
+		"SpecialLoopExport": "includes/LoopExport.php",
 		"LoopSettings": "includes/LoopSettings.php",
-		"SpecialLoopSettings": "includes/SpecialLoopSettings.php",
-		"SpecialPurgeCache": "includes/SpecialPurgeCache.php",
+		"SpecialLoopSettings": "includes/LoopSettings.php",
+		"SpecialPurgeCache": "includes/PurgeCache.php",
 		"LoopException": "includes/LoopException.php",
 		"LoopObjectIndex": "includes/LoopObjectIndex.php",
 		"LoopObject": "includes/LoopObject.php",
@@ -65,9 +65,9 @@
 		"LoopReference": "includes/LoopReference.php",
 		"LoopCode": "includes/LoopCode.php",
 		"LoopLiterature": "includes/LoopLiterature.php",
-		"SpecialLoopLiterature": "includes/SpecialLoopLiterature.php",
-		"SpecialLoopLiteratureEdit": "includes/SpecialLoopLiterature.php",
-		"SpecialLoopLiteratureImport": "includes/SpecialLoopLiterature.php",
+		"SpecialLoopLiterature": "includes/LoopLiterature.php",
+		"SpecialLoopLiteratureEdit": "includes/LoopLiterature.php",
+		"SpecialLoopLiteratureImport": "includes/LoopLiterature.php",
 		"LoopGlossary": "includes/LoopGlossary.php",
 		"SpecialLoopGlossary": "includes/LoopGlossary.php"
 	},
